sonarcloud bug fixes

+++ Source/CDR.DiscoverDataHolders/DiscoverDataHolders.cs
@@ -254,16 +254,20 @@ namespace CDR.DiscoverDataHolders
             }
 
             log.LogInformation("Synchronising existing {count} data holder brands", existingBrands?.Count());
-            foreach (var existingDataHolderBrand in existingBrands)
-            {                
-                //existing data holders that don't exist in mdh should be removed from the mdr
-                var exists = data.Any(x => x.DataHolderBrandId.Equals(existingDataHolderBrand.DataHolderBrandId, StringComparison.OrdinalIgnoreCase));
-                
-                //Remove additional or extra brands to reflect correct brand data
-                if (!exists)
+
+            if (existingBrands.Any())
+            {
+                foreach (var existingDataHolderBrand in existingBrands)
                 {
-                    log.LogInformation("Deleting existing data holder brand: {brandId}", existingDataHolderBrand.DataHolderBrandId);
-                    await sql.DeleteDataHolder(existingDataHolderBrand.DataHolderBrandId);                    
+                    //existing data holders that don't exist in mdh should be removed from the mdr
+                    var exists = data.Any(x => x.DataHolderBrandId.Equals(existingDataHolderBrand.DataHolderBrandId, StringComparison.OrdinalIgnoreCase));
+
+                    //Remove additional or extra brands to reflect correct brand data
+                    if (!exists)
+                    {
+                        log.LogInformation("Deleting existing data holder brand: {brandId}", existingDataHolderBrand.DataHolderBrandId);
+                        await sql.DeleteDataHolder(existingDataHolderBrand.DataHolderBrandId);
+                    }
                 }
             }
         }
